+++ src/core/Sampler/Sampler.cpp
@@ -1042,14 +1042,9 @@ bool Sampler::processPlaybackTrack(int nBufferSize)
 	}
 
 	const auto pCompo = m_pPlaybackTrackInstrument->getComponents()->front();
-	if ( pCompo == nullptr ) {
-		ERRORLOG( "Invalid component of playback instrument" );
-		return true;
-	}
-
-	auto pSample = pCompo->getLayer(0)->getSample();
-	if ( pSample == nullptr ) {
-		ERRORLOG( "Unable to process playback track" );
+	if ( pCompo == nullptr || pCompo->getLayer( 0 ) == nullptr ||
+		pCompo->getLayer( 0 )->getSample() == nullptr ) {
+		ERRORLOG( "Invalid playback instrument" );
 		EventQueue::get_instance()->pushEvent( Event::Type::Error,
 												Hydrogen::ErrorMessages::PLAYBACK_TRACK_INVALID );
 		// Disable the playback track
@@ -1058,13 +1053,16 @@ bool Sampler::processPlaybackTrack(int nBufferSize)
 		return true;
 	}
 
+	auto pSample = pCompo->getLayer( 0 )->getSample();
+
 	auto pSample_data_L = pSample->getData_L();
 	auto pSample_data_R = pSample->getData_R();
 
 	int nAvail_bytes = 0;
-	int	nInitialBufferPos = 0;
+	int nInitialBufferPos = 0;
 
-	const long long nFrame = pAudioEngine->getTransportPosition()->getFrame();
+	const long long nFrame = pAudioEngine->getTransportPosition()->getFrame() -
+							 pAudioEngine->getLastLoopFrame();
 	const long long nFrameOffset =
 		pAudioEngine->getTransportPosition()->getFrameOffsetTempo();
 
@@ -1076,6 +1074,10 @@ bool Sampler::processPlaybackTrack(int nBufferSize)
 							 nBufferSize );
 
 	int nFinalBufferPos = nInitialBufferPos + nAvail_bytes;
+	if ( nFinalBufferPos < 0 ) {
+		// end of sample reached.
+		return true;
+	}
 
 	// Output-rate buffer as temporary storage for sample data, resampled to output rate
 	float buffer_L[ nBufferSize ];
@@ -1517,7 +1519,7 @@ void Sampler::reinitializePlaybackTrack()
 		pSample = Sample::load( pSong->getPlaybackTrackFilename() );
 	}
 	
-	auto  pPlaybackTrackLayer = std::make_shared<InstrumentLayer>( pSample );
+	auto pPlaybackTrackLayer = std::make_shared<InstrumentLayer>( pSample );
 
 	m_pPlaybackTrackInstrument->getComponents()->front()->setLayer( pPlaybackTrackLayer, 0 );
 	m_nPlayBackSamplePosition = 0;